Use set +/-x and test

+++ travis/travis.sh
@@ -19,18 +19,27 @@ export XML_CATALOG_FILES="catalog.xml"
 
 # -----------------------------------------------------------
 function travis_fold_start() {
+  set +x
   echo -e "travis_fold:start:$1${YELLOW}$2${NC}"
+  set -x
 }
 
 function travis_fold_end() {
+  set +x
   echo -e "\ntravis_fold:end:$1\r"
+  set -x
 }
 
-set -x
+# set -x
 # -----------------------------------------------------------
 echo "Using variables:
   CACHE='$CACHE'
   XML_CATALOG_FILES='$XML_CATALOG_FILES'"
+
+# -----------------------------------------------------------
+travis_fold_start cache "List cache"
+ls $CACHE
+travis_fold_end cache
 
 # -----------------------------------------------------------
 travis_fold_start catalog_creation "Create XML catalog"
@@ -49,15 +58,14 @@ travis_fold_end catalog_check
 
 # ---------------------
 travis_fold_start download "Download and cache"
-echo -en 'travis_fold:start:download\r'
 test -e $CACHE || ( mkdir -p $CACHE; wget -P $CACHE http://docbook.org/xml/4.5/docbook-xml-4.5.zip; unzip -d $CACHE $CACHE/docbook-xml-4.5.zip )
 travis_fold_end download
 
 
 # ---------------------
 travis_fold_start entities "Copy entities"
-cp -vi $CACHE/*.mod xml/
-cp -avi $CACHE/ent xml/
+test -e $CACHE/*.mod && cp -vi $CACHE/*.mod xml/
+test -e $CACHE/ent   && cp -avi $CACHE/ent xml/
 travis_fold_end entities
